requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.1.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dlib==19.24.4
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-MySQLdb==2.0.0
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
gast==0.5.4
gevent==24.2.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
idna==3.7
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mtcnn==0.1.1
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
Pygments==2.18.0
PyJWT==2.8.0
PyMySQL==1.1.1
python-engineio==4.9.1
python-socketio==5.11.2
requests==2.32.3
rich==13.7.1
simple-websocket==1.0.0
six==1.16.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
typing_extensions==4.12.1
urllib3==1.26.6
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
zipp==3.19.0
zope.event==5.0
zope.interface==6.4.post2
